Split out Value into a separate crate: cranelift-value
This is necessary to avoid a circular dependency between cranelift-filetests and cranelift-interpreter. If in the future FunctionRunner is moved from cranelift-filetests to cranelift-interpreter (which it probably should be), then cranelift-value can be collapsed again into cranelift-interpreter. (This also brings in some cranelift-filetests imports that should probably be in the next commit).